pyproject.toml
[tool.poetry]
name = "nonebot-plugin-cp-broadcast"
version = "2.3.0"
description = "Codeforces、牛客、AtCoder平台比赛查询，ACMer 必备"
authors = ["HuParry <hu_parry@qq.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "nonebot_plugin_cp_broadcast"}]
homepage = "https://github.com/HuParry/nonebot-plugin-cp-broadcast"
repository = "https://github.com/HuParry/nonebot-plugin-cp-broadcast"

[tool.poetry.dependencies]
python = ">=3.8"
nonebot2 = { version = "^2.0.0", extras = ["fastapi"] }
nonebot-adapter-onebot = ">=2.1.5"
httpx = ">=0.20.0,<1.0.0"
fake-useragent = ">=1.1.3"
aiosqlite = "^0.13.0"
bs4 = ">=0.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
